Skeletal view feedback fixes
